requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
langchain
python-dotenv
openai
streamlit
requests
pytest
